change TOML_TYPE.TRUE/FALSE to BOOL
This is a breaking change.

this is so that you can more easily check for boolean types at runtime
without a special case for booleans to check both true and false.

true and false as separate types makes it so it's not neccessary to
check any other memory (the union), but in practice the union is
already the same size and checking for a boolean always becomes a 2-step
operation to check if it is either true or false, otherwise throw an
exception.